+++ crates/semantic_index/src/semantic_index.rs
@@ -1,4 +1,7 @@
+mod chunking;
+
 use anyhow::{anyhow, Context as _, Result};
+use chunking::{chunk_text, Chunk};
 use client::Client;
 use collections::HashMap;
 use fs::Fs;
@@ -140,18 +143,25 @@ impl ProjectIndex {
                     for _ in 0..cx.num_cpus() {
                         cx.spawn(async {
                             while let Ok(entry) = entries.recv().await {
-                                if let Some(chunked_file) = chunk_file(
-                                    worktree_abs_path.clone(),
+                                let entry_abs_path = worktree_abs_path.join(&entry.path);
+                                let Some(text) = fs.load(&entry_abs_path).await.log_err() else {
+                                    continue;
+                                };
+                                let language = language_registry
+                                    .language_for_file_path(&entry.path)
+                                    .await
+                                    .ok();
+                                let grammar =
+                                    language.as_ref().and_then(|language| language.grammar());
+                                let chunked_file = ChunkedFile {
+                                    worktree_root: worktree_abs_path.clone(),
+                                    chunks: chunk_text(&text, grammar),
                                     entry,
-                                    &language_registry,
-                                    fs.as_ref(),
-                                )
-                                .await
-                                .log_err()
-                                {
-                                    if chunked_files_tx.send(chunked_file).await.is_err() {
-                                        break;
-                                    }
+                                    text,
+                                };
+
+                                if chunked_files_tx.send(chunked_file).await.is_err() {
+                                    break;
                                 }
                             }
                         });
@@ -167,6 +177,7 @@ impl ProjectIndex {
         chunked_files: channel::Receiver<ChunkedFile>,
         cx: &mut ModelContext<ProjectIndex>,
     ) -> Task<()> {
+        let http_client = self.client.http_client();
         cx.spawn(|this, cx| async move {
             let mut chunked_file_batches =
                 chunked_files.chunks_timeout(512, Duration::from_secs(2));
@@ -177,161 +188,17 @@ impl ProjectIndex {
     }
 }
 
-impl EventEmitter<StatusEvent> for ProjectIndex {}
-
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
 pub enum StatusEvent {
     Idle,
     Scanning,
 }
 
+impl EventEmitter<StatusEvent> for ProjectIndex {}
+
 struct ChunkedFile {
     worktree_root: Arc<Path>,
     entry: Entry,
     text: String,
     chunks: Vec<Chunk>,
-}
-
-struct Chunk {
-    range: Range<usize>,
-    digest: [u8; 32],
-}
-
-async fn chunk_file(
-    worktree_root: Arc<Path>,
-    entry: Entry,
-    language_registry: &Arc<LanguageRegistry>,
-    fs: &dyn Fs,
-) -> Result<ChunkedFile> {
-    let abs_path = worktree_root.join(&entry.path);
-
-    let text = fs.load(&abs_path).await?;
-    let language = language_registry
-        .language_for_file_path(&entry.path)
-        .await
-        .ok();
-
-    let chunks = if let Some(grammar) = language.as_ref().and_then(|language| language.grammar()) {
-        let tree = PARSER.with(|parser| {
-            let mut parser = parser.borrow_mut();
-            parser
-                .set_language(&grammar.ts_language)
-                .expect("incompatible grammar");
-            parser.parse(&text, None).expect("invalid language")
-        });
-
-        chunk_parse_tree(tree, &text)
-    } else {
-        chunk_lines(&text)
-    };
-
-    Ok(ChunkedFile {
-        worktree_root,
-        entry,
-        text,
-        chunks,
-    })
-}
-
-const CHUNK_THRESHOLD: usize = 1500;
-
-fn chunk_parse_tree(tree: Tree, text: &str) -> Vec<Chunk> {
-    let mut chunk_ranges = Vec::new();
-    let mut cursor = tree.walk();
-
-    let mut range = 0..0;
-    loop {
-        let node = cursor.node();
-
-        // If adding the node to the current chunk exceeds the threshold
-        if node.end_byte() - range.start > CHUNK_THRESHOLD {
-            // Try to descend into its first child. If we can't, flush the current
-            // range and try again.
-            if cursor.goto_first_child() {
-                continue;
-            } else if !range.is_empty() {
-                chunk_ranges.push(range.clone());
-                range.start = range.end;
-                continue;
-            }
-
-            // If we get here, the node itself has no children but is larger than the threshold.
-            // Break its text into arbitrary chunks.
-            chunk_text(text, range.clone(), node.end_byte(), &mut chunk_ranges);
-        }
-        range.end = node.end_byte();
-
-        // If we get here, we consumed the node. Advance to the next child, ascending if there isn't one.
-        while !cursor.goto_next_sibling() {
-            if !cursor.goto_parent() {
-                if !range.is_empty() {
-                    chunk_ranges.push(range);
-                }
-
-                return chunk_ranges
-                    .into_iter()
-                    .map(|range| {
-                        let mut hasher = Sha256::new();
-                        hasher.update(&text[range.clone()]);
-                        let mut digest = [0u8; 32];
-                        digest.copy_from_slice(hasher.finalize().as_slice());
-                        Chunk { range, digest }
-                    })
-                    .collect();
-            }
-        }
-    }
-}
-
-fn chunk_lines(text: &str) -> Vec<Chunk> {
-    let mut chunk_ranges = Vec::new();
-    let mut range = 0..0;
-
-    let mut newlines = text.match_indices('\n').peekable();
-    while let Some((newline_ix, _)) = newlines.peek() {
-        let newline_ix = newline_ix + 1;
-        if newline_ix - range.start <= CHUNK_THRESHOLD {
-            range.end = newline_ix;
-            newlines.next();
-        } else {
-            if range.is_empty() {
-                chunk_text(text, range, newline_ix, &mut chunk_ranges);
-                range = newline_ix..newline_ix;
-            } else {
-                chunk_ranges.push(range.clone());
-                range.start = range.end;
-            }
-        }
-    }
-
-    if !range.is_empty() {
-        chunk_ranges.push(range);
-    }
-
-    chunk_ranges
-        .into_iter()
-        .map(|range| {
-            let mut hasher = Sha256::new();
-            hasher.update(&text[range.clone()]);
-            let mut digest = [0u8; 32];
-            digest.copy_from_slice(hasher.finalize().as_slice());
-            Chunk { range, digest }
-        })
-        .collect()
-}
-
-fn chunk_text(
-    text: &str,
-    mut range: Range<usize>,
-    max_end: usize,
-    chunk_ranges: &mut Vec<Range<usize>>,
-) {
-    while range.start < max_end {
-        range.end = cmp::min(range.start + CHUNK_THRESHOLD, max_end);
-        while !text.is_char_boundary(range.end) {
-            range.end -= 1;
-        }
-        chunk_ranges.push(range.clone());
-        range.start = range.end;
-    }
 }
